feat: Add BF16 support gemm in moore gpu and rename existing 'musa' to 'moore' in some files
