+++ src/infiniop/ops/gemm/moore/gemm_moore.h
@@ -1,0 +1,8 @@
+#ifndef __GEMM_MOORE_H__
+#define __GEMM_MOORE_H__
+
+#include "../gemm.h"
+
+DESCRIPTOR(moore)
+
+#endif // __GEMM_MOORE_H__
