--- src/infiniop/ops/gemm/musa/gemm_musa.h
@@ -1,8 +1,0 @@
-#ifndef __GEMM_MUSA_H__
-#define __GEMM_MUSA_H__
-
-#include "../gemm.h"
-
-DESCRIPTOR(musa)
-
-#endif // __GEMM_MUSA_H__
