+++ src/infiniop/ops/rms_norm/moore/rms_norm_moore.h
@@ -1,0 +1,8 @@
+#ifndef __RMS_NORM_MOORE_H__
+#define __RMS_NORM_MOORE_H__
+
+#include "../rms_norm.h"
+
+DESCRIPTOR(moore)
+
+#endif
