--- src/infiniop/ops/rms_norm/musa/rms_norm_musa.h
@@ -1,8 +1,0 @@
-#ifndef __RMS_NORM_MUSA_H__
-#define __RMS_NORM_MUSA_H__
-
-#include "../rms_norm.h"
-
-DESCRIPTOR(musa)
-
-#endif
