+++ xmake/moore.lua
@@ -42,8 +42,8 @@ target("infiniop-moore")
     set_languages("cxx17")
     set_warnings("all", "error")
     add_cxflags("-lstdc++", "-fPIC", "-Wno-comment")
-    add_files("../src/infiniop/devices/musa/*.cc")
-    add_files("../src/infiniop/ops/*/musa/*.mu", {rule = "mu"})
+    add_files("../src/infiniop/devices/moore/*.cc")
+    add_files("../src/infiniop/ops/*/moore/*.mu", {rule = "mu"})
 target_end()
 
 target("infinirt-moore")
@@ -53,5 +53,5 @@ target("infinirt-moore")
     add_deps("infini-utils")
     set_warnings("all", "error")
     add_cxflags("-lstdc++", "-fPIC")
-    add_files("../src/infinirt/musa/*.cc")
+    add_files("../src/infinirt/moore/*.cc")
 target_end()
